fix: lazy import preload, then update include.type to include.as
