+++ .changeset/blue-pears-allow.md
@@ -1,0 +1,7 @@
+---
+'@modern-js/prod-server': patch
+'@modern-js/server-core': patch
+---
+
+fix: lazy import preload, then update include.type to include.as
+fix: 动态加载 preload 功能, 对齐规范将 include.type 属性更新为 include.as
