+++ packages/server/prod-server/src/libs/preload/flushServerHeader.ts
@@ -4,41 +4,6 @@ import { ModernServerContext } from '@modern-js/types';
 import { parseLinks } from './parseLinks';
 import { transformLinks2String } from './transformLinks2String';
 
-export function transformToRegExp(input: string | RegExp): RegExp {
-  if (typeof input === 'string') {
-    return new RegExp(input);
-  }
-  return input;
-}
-
-export function shouldFlushServerHeader(
-  serverConf: ServerOptions['server'],
-  userAgent?: string,
-  disablePreload?: boolean,
-) {
-  const { ssr: ssrConf } = serverConf || {};
-
-  if (disablePreload) {
-    return false;
-  }
-
-  if (typeof ssrConf === 'object' && ssrConf.preload) {
-    // ssr.preload: 'object'
-    if (typeof ssrConf.preload === 'object') {
-      const { userAgentFilter } = ssrConf.preload;
-      if (userAgentFilter && userAgent) {
-        return !transformToRegExp(userAgentFilter).test(userAgent);
-      }
-      return true;
-    }
-    // ssr.preload: true;
-    return true;
-  }
-
-  // ssr: false or ssr: true
-  return false;
-}
-
 export interface FlushServerHeaderOptions {
   ctx: ModernServerContext;
   distDir: string;
