+++ tests/integration/ssr/fixtures/preload/modern.config.ts
@@ -3,7 +3,14 @@ import { appTools, defineConfig } from '@modern-js/app-tools';
 export default defineConfig({
   server: {
     ssr: {
-      preload: true,
+      preload: {
+        include: [
+          {
+            url: 'https://lf3-static.bytednsdoc.com/obj/eden-cn/nuvshpqnulg/eden-x-logo.png',
+            as: 'image',
+          },
+        ],
+      },
     },
   },
   runtime: {
